pyproject.toml
[project]
name = "ai-agent-system"
version = "1.0.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic-settings>=2.7",
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "langchain-postgres>=0.0.13",
    "langgraph-checkpoint-redis>=0.0.5",
    "langchain-community>=0.3",
    "pypdf>=5.1",
    "beautifulsoup4>=4.12",
    "redis>=5.2",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "structlog>=24.4",
    "httpx>=0.28",
    "python-jose[cryptography]>=3.3",
    "passlib[bcrypt]>=1.7",
    "langfuse>=2.50",
    "arq>=0.26",
    "prometheus-fastapi-instrumentator>=7.0",
    "sqlparse>=0.5",
    "tenacity>=9.0",
    "uvloop>=0.20",
]

[tool.uv]
dev-dependencies = ["pytest>=8.3", "pytest-asyncio>=0.24", "httpx>=0.28"]